blame is run in the background as a job
Instead of running blame as the cursor moves, a single big
git blame is executed when things blamer is started for the
first time.  Once that blame completes, the results are
parsed and pushed into two tables
  *        line -> commit-hash
  * commit-hash -> commit-information

For large git repos with long histories, the git blame operation
can take a long time and doing it this way keeps nvim usable
while git figures things out.  And the long operation doesn't have
to be done over and over again as you move around the file.

Still to do:
* information doesn't get updated.  Ever.  Even as you edit the file.
* no handling for something that isn't a git file
* editing a git file in another repo isn't handled correctly (it wasn't
  before either).  This should be easy to fix..
* data is not kept 'buffer local' so the new file just sees the same
  git blame data as the original file -- how do you do buffer local data
  with lua?