automatic buffer local disabling of blamer
Used for files that can't be blamed (not in a repo, or not part of the
repo).  That check is then only run once before we launch a blame job
and if it turns out that the file can't be blamed, then the buffer local
flag is disabled.

All buffer local variables are prefixed with blamer to minimzie the
changes of conflicting with someone else's flags.